// resources/js/api/generated/models/getApiV1AdminSecurityAuditLogs200DataLogsItem.ts
/**
 * Generated by orval v8.0.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */

export type GetApiV1AdminSecurityAuditLogs200DataLogsItem = {
  actor_ip: string;
  actor_user_agent: string;
  /** @minimum 0 */
  actor_user_id: number;
  actor_username: string;
  created_at: string;
  event_category: string;
  event_type: string;
  failure_reason: string;
  /** @minimum 0 */
  id: number;
  metadata: string;
  /** @minimum 0 */
  server_id: number;
  session_id: string;
  severity: string;
  stack_name: string;
  success: boolean;
  /** @minimum 0 */
  target_id: number;
  target_name: string;
  target_type: string;
  /** @minimum 0 */
  target_user_id: number;
  updated_at: string;
};